requirements.txt
aws-cdk-lib>=2.92.0
constructs>=10.0.0,<11.0.0

pip>=22.0.4
cryptography~=37.0.2
setuptools~=60.0.3
future~=0.18.2
docutils~=0.15.2
boto3>=1.19.0
